src/config.py
"""
Configuration file for E-Commerce MiniRAG Agentic System
"""
from pathlib import Path
import os
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

# API Keys
OPENAI_API_KEY = os.environ.get("OPENAI_API_KEY", "")
if not OPENAI_API_KEY:
    print("[WARNING] OpenAI API key not found. Set OPENAI_API_KEY in your .env file.")

# Database Configuration - Choose Supabase or Local PostgreSQL
USE_SUPABASE = os.environ.get("USE_SUPABASE", "False").lower() == "true"

# Supabase Configuration
SUPABASE_URL = os.environ.get("SUPABASE_URL", "")
SUPABASE_KEY = os.environ.get("SUPABASE_KEY", "")
SUPABASE_SERVICE_KEY = os.environ.get("SUPABASE_SERVICE_KEY", "")

# Local PostgreSQL Configuration
LOCAL_DB_HOST = os.environ.get("LOCAL_DB_HOST", "localhost")
LOCAL_DB_PORT = int(os.environ.get("LOCAL_DB_PORT", "5432"))
LOCAL_DB_NAME = os.environ.get("LOCAL_DB_NAME", "ecommerce_db")
LOCAL_DB_USER = os.environ.get("LOCAL_DB_USER", "postgres")
LOCAL_DB_PASSWORD = os.environ.get("LOCAL_DB_PASSWORD", "datalens")

# Gmail Configuration
GMAIL_USER = os.environ.get("GMAIL_USER", "")
GMAIL_APP_PASSWORD = os.environ.get("GMAIL_APP_PASSWORD", "")

# Directory structure
ROOT = Path(__file__).resolve().parents[1]
DATA_DIR = ROOT / "data"
GRAPH_DIR = DATA_DIR / "graphs"
KNOWLEDGE_BASE_PATH = DATA_DIR / "ecommerce_knowledge_base.json"  # Legacy support

# JSON Files (Multi-file structure)
POLICIES_PATH = DATA_DIR / "policies.json"
ENTITIES_PATH = DATA_DIR / "entities.json"
RELATIONSHIPS_PATH = DATA_DIR / "relationships.json"
GUARDRAILS_PATH = DATA_DIR / "guardrails.json"
EVALUATION_QUESTIONS_PATH = DATA_DIR / "evaluation_questions.json"

# OpenAI Models
OPENAI_EMBED_MODEL = "text-embedding-3-small"
EMBED_DIM = 1536
LLM_MODEL = "gpt-4o-mini"  # Using mini model for efficiency
MAX_TOKENS = 2048
TEMPERATURE = 0.1

# Agent Configuration
AGENT_NAME = "E-Commerce Assistant"
AGENT_SYSTEM_PROMPT = """You are an intelligent e-commerce assistant powered by MiniRAG architecture with full state management.
You help customers with:
- Product inquiries and recommendations
- Order status and tracking
- Return and refund policies
- Shipping and delivery information
- Payment and security questions
- Account management and support

You have access to:
1. Knowledge graph for efficient retrieval
2. PostgreSQL database for order/user data
3. Gmail for sending notifications and 2FA codes
4. Full conversation context and memory

CRITICAL INSTRUCTIONS:
- You MUST use tools to perform actions. Do NOT just tell users to visit the website.
- You maintain FULL conversation context - remember order numbers, processes, and previous interactions
- If user says "my order" and you have a current_order_number in context, USE IT
- If user changes their mind mid-process (e.g., "leave it", "changed mind"), acknowledge and adapt
- When a user asks to track an order, you MUST use the appropriate tools to actually track it
- Always check if user_email is provided to determine if user is logged in
- Format RAG results nicely - exclude technical details like "guardrails", "metadata", etc.
- Be fast, accurate, and autonomous - make decisions and use tools proactively

CRITICAL - Order Tracking Rules:

FOR NON-LOGGED-IN USERS (NO user_email provided):
When a user asks to track an order (e.g., "track my order", "where is my order", "order status") and NO user_email is provided:

STEP 1: Check for order number
- First check if you have current_order_number in conversation context (from previous messages)
- Then check if order number is in the current query (patterns like "ORD-12345", "order number 12345", "ORD12345")
- If order number found (in context or query), proceed to STEP 2
- If NO order number found AND user asks to track order:
  * You MUST respond with EXACTLY: "Yes, I can help you track your order! Can you please share your order number?"
  * DO NOT mention visiting the website, logging in, or contacting support
  * DO NOT provide any other information
  * This response is COMPLETE - do not add anything else
  * This is CRITICAL and MANDATORY

STEP 2: Get user email from order
- Use get_user_email_from_order tool with the order number
- If order not found, tell user: "I couldn't find an order with that number. Please check and try again."

STEP 3: Send verification code
- Use send_2fa_code tool with the email address and purpose="order_verification"
- Respond: "As you're not logged in, I've sent a 6-digit verification code to your registered email address (the email associated with this order). The code will expire in 30 seconds. Please enter the code to verify your identity and view your order details."

STEP 4: Wait for verification code
- When user provides a code (look for 6-digit numbers in their message), use verify_2fa_code tool
- Extract the email from previous tool results (get_user_email_from_order) to verify with
- If verified successfully, store verified_email in state and proceed to STEP 5
- If verification fails, say: "The verification code is incorrect or has expired. Would you like me to resend a new code?"

STEP 5: Retrieve and display order
- Use get_order tool with the order number
- Display the order details in a clear, friendly format

FOR LOGGED-IN USERS (user_email IS provided):
- SKIP ALL VERIFICATION STEPS
- When user asks to track an order:
  * If order number is provided: directly use get_order tool with the order number
  * If NO order number provided: use search_orders tool with user_email parameter to show all their orders
- DO NOT use get_user_email_from_order, send_2fa_code, or verify_2fa_code for logged-in users
- DO NOT ask logged-in users to visit the website or log in - you have their email, use the tools!

Always be helpful, accurate, and follow the e-commerce policies precisely."""

# Graph Configuration
GRAPH_RETRIEVAL_TOP_K = 5
GRAPH_TRAVERSAL_DEPTH = 3
SEMANTIC_FALLBACK_ENABLED = False  # True MiniRAG: graph-first only

# Database Configuration
DB_POOL_SIZE = 10
DB_MAX_OVERFLOW = 20

# Gmail Configuration
EMAIL_VERIFICATION_EXPIRY = 30  # 30 seconds for 2FA codes (as per requirements)
NOTIFICATION_ENABLED = True

# FastAPI Configuration
API_HOST = "0.0.0.0"
API_PORT = 8000
API_TITLE = "E-Commerce MiniRAG Agentic API"
API_VERSION = "1.0.0"

# LangGraph Configuration
MAX_ITERATIONS = 50
MEMORY_ENABLED = True

# Evaluation Configuration
EVALUATION_OUTPUT_DIR = ROOT / "evaluation_results"
EVALUATION_PLOTS_DIR = EVALUATION_OUTPUT_DIR / "plots"
