requirements.txt
# Core Dependencies
openai>=1.10.0
python-dotenv>=1.0.0

# FastAPI and Web Server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# LangGraph and LangChain
langgraph>=0.0.20
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-core>=0.1.0

# Graph and Network
networkx>=3.1
faiss-cpu>=1.7.4

# Database and Tools
supabase>=2.0.0
psycopg2-binary>=2.9.9

# Email
secure-smtplib>=0.1.1

# Streamlit
streamlit>=1.26.0

# Utilities
tqdm>=4.66.1
numpy>=1.23.0
pandas>=1.5.0

# Evaluation and Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# MCP (Model Context Protocol)
mcp>=0.9.0

# Development
pytest>=7.4.0
black>=23.11.0
